cast to float when loading numpy array

+++ src/datamodules/era5_iterdataset.py
@@ -119,9 +119,9 @@ class ERA5Forecast(IterableDataset):
         # i.e. where inputs are from previous years and output from next
         for data, variables, out_variables in self.dataset:
             min_len = np.min([data[k].shape[0] for k in data.keys()])
-            x = np.concatenate([data[k][:min_len] for k in data.keys()], axis=1)
+            x = np.concatenate([data[k][:min_len].astype(np.float32) for k in data.keys()], axis=1)
             x = torch.from_numpy(x)
-            y = np.concatenate([data[k][:min_len] for k in out_variables], axis=1)
+            y = np.concatenate([data[k][:min_len].astype(np.float32) for k in out_variables], axis=1)
             y = torch.from_numpy(y)
 
             inputs = x.unsqueeze(0).repeat_interleave(self.history, dim=0)
@@ -154,9 +154,9 @@ class ERA5ForecastMultiStep(IterableDataset):
         # i.e. where inputs are from previous years and output from next
         for data, variables, out_variables in self.dataset:
             min_len = np.min([data[k].shape[0] for k in data.keys()])
-            x = np.concatenate([data[k][:min_len] for k in data.keys()], axis=1)
+            x = np.concatenate([data[k][:min_len].astype(np.float32) for k in data.keys()], axis=1)
             x = torch.from_numpy(x)
-            y = np.concatenate([data[k][:min_len] for k in out_variables], axis=1)
+            y = np.concatenate([data[k][:min_len].astype(np.float32) for k in out_variables], axis=1)
             y = torch.from_numpy(y)
 
             inputs = x.unsqueeze(0).repeat_interleave(self.history, dim=0)
@@ -189,7 +189,7 @@ class IndividualForecastDataIter(IterableDataset):
             for i in range(inp.shape[0]):
                 # TODO: should we unsqueeze the first dimension?
                 if self.transforms is not None:
-                    yield self.transforms(inp[i]).to(torch.float), self.output_transforms(out[i]).to(torch.float), variables, out_variables
+                    yield self.transforms(inp[i]), self.output_transforms(out[i]), variables, out_variables
                 else:
                     yield inp[i], out[i], variables, out_variables
